Issue #568: Get custom home tile titles synchronously.
In order to remove the animations, we need to retrieve the titles and
screenshots synchronously: this is the first step.

The more complex solution I mention in the comments would have us store
a single Deferred as the return value of `PublicSuffixPatterns.getExactSet`,
which will only need to be completed once. However, it requires chaining many
returned `Deferred`s, which has a lot of complexity:
- We need to ensure we're dispatching on the right thread (i.e. synchronously
if the deferred is completed)
- If it's cancelled (e.g. automatically by the UI thread), we need to
rerun the job at a later point
- We need to handle exceptions
- Does a completed coroutine hold strong references to the variables its
function block captured? (I hope not!)

So I opted to do the simple thing and block the UI thread only on first
run.