Issue #1188: Add danger type definitions; update licenses accordingly.
The danger type definitions are licensed under the MIT license to match
the license distributed by danger in order to facilitate re-use: I spoke with
Mike Hoye who mentioned this is a standard thing for us to do. I updated
the other license files in the repository as appropriate.

One day we may wish to auto-generate these files - e.g. with ts2kt, but
this is working now and that isn't. :)